requirements.txt
flask
flask_mail
flask-cors
python-dotenv
